+++ README.md
@@ -3,3 +3,15 @@
 This template should help get you started developing with Vue 3 in Vite. The template uses Vue 3 `<script setup>` SFCs, check out the [script setup docs](https://v3.vuejs.org/api/sfc-script-setup.html#sfc-script-setup) to learn more.
 
 Learn more about IDE Support for Vue in the [Vue Docs Scaling up Guide](https://vuejs.org/guide/scaling-up/tooling.html#ide-support).
+
+https://icones.js.org/collection/lucide
+
+```html
+  
+        <h3> remarque importante </h3> 
+        <p> Pour une écoute sans interruption, je vous recommande fortement de procéder comme suit: 
+            1) télécharger l'enregistrement choisi à l'aide du bouton à droite de celui-ci. 
+            2) mettre votre appareil en mode avion 
+            3) appuyer sur play! En cas de problème, les enregistrements sont aussi disponibles à l'écoute et au téléchargement au format mp3 <a href='https://www.enpleineconscience.ch/meditations-guidees-mp3/'>sur mon site</a>. 
+            </p>'
+```
