+++ package.json
@@ -9,13 +9,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "tailwindcss": "^4.1.11",
     "vue": "^3.5.13",
     "vue-router": "^4.5.1"
   },
   "devDependencies": {
+    "@tailwindcss/vite": "^4.1.11",
     "@vitejs/plugin-vue": "^5.2.3",
     "vite": "^6.3.5",
-    "vite-plugin-pwa": "^1.0.0"
+    "vite-plugin-pwa": "^1.0.0",
+    "vite-svg-loader": "^5.1.0"
   },
   "imports": {
     "#src/*": "./src/*"
