+++ src/_/routes/index.vue
