+++ src/components/l-download.vue
@@ -1,0 +1,25 @@
+<!--
+    Template
+-->
+
+<template>
+  <aside class="l-download fixed w-full py-4 bottom-[0] left-[0] bg-white shadow-(--shadow)">
+    <div class="wrapper h-full">
+      <p>Downloading 1 item(s)</p>
+      <p class="text-sm text-slate-600 mb-4">0.1 MB of 120.7 MB</p>
+      <div class="h-1 bg-slate-200">
+        <div class="h-full bg-blue-500 w-[33%]" />
+      </div>
+    </div>
+  </aside>
+</template>
+
+
+<!--
+    Scripts
+-->
+
+<script setup lang="ts">
+
+
+</script>
