+++ src/config/categories/kindness.js
@@ -1,6 +1,7 @@
 export default {
     name: 'bienveillance',
-    description: 'La méditation de bienveillance se pratique dans une position assise confortable | et consiste à répéter intérieurement une phrase exprimant une intention bienveillante en l\'adressant à soi-même et à d\'autres personnes. Elle est présentée dans le détail dans le cadre du cours. A chacun de choisir la phrase qui convient le mieux, à condition d\'utilliser la même pour les différentes personnes auxquelles vous penserez et pour vous-même.  Plusieurs phrases sont proposées dans le cadre des formations, l\'une des plus utilisées étant par exemple \"Je te souhaite d\'être heureux et  en bonne santé.\" Certains des exercices ci-dessous intègrent une phase guidée de choix de la phrase utilisée. Il y aussi des enregistrement courts dans la catégorie \"mini-assises\".  <h3> remarque importante </h3> <p> Pour une écoute sans interruption, je vous recommande fortement de procéder comme suit: 1) télécharger l\'enregistrement choisi à l\'aide du bouton à droite de celui-ci. 2) mettre votre appareil en mode avion 3) appuyer sur play! En cas de problème, les enregistrements sont aussi disponibles à l\'écoute et au téléchargement au format mp3 <a href=\"https://www.enpleineconscience.ch/meditations-guidees-mp3/\">sur mon site</a>. </p>',
+    slug: 'bienveillance',
+    description: 'La méditation de bienveillance se pratique dans une position assise confortable | et consiste à répéter intérieurement une phrase exprimant une intention bienveillante en l\'adressant à soi-même et à d\'autres personnes. Elle est présentée dans le détail dans le cadre du cours. A chacun de choisir la phrase qui convient le mieux, à condition d\'utilliser la même pour les différentes personnes auxquelles vous penserez et pour vous-même.  Plusieurs phrases sont proposées dans le cadre des formations, l\'une des plus utilisées étant par exemple \"Je te souhaite d\'être heureux et  en bonne santé.\" Certains des exercices ci-dessous intègrent une phase guidée de choix de la phrase utilisée. Il y aussi des enregistrement courts dans la catégorie \"mini-assises\".',
     recordings: [
         {
             name: 'bienveillance, 34\' (4/17)',
