+++ src/config/categories/silence.js
@@ -1,6 +1,7 @@
 export default {
     name: 'silence',
-    description: 'Cet enregistrement de silence avec des sons de bol permet de méditer sans instructions. 1 \"ding\" après 10 minutes, 2 après 20, 3 après 30, etc. jusqu\'à 6 coups pour 60 minutes.  <h3> remarque importante </h3> <p> Pour une écoute sans interruption, je vous recommande fortement de procéder comme suit: 1) télécharger l\'enregistrement choisi à l\'aide du bouton à droite de celui-ci. 2) mettre votre appareil en mode avion 3) appuyer sur play! En cas de problème, les enregistrements sont aussi disponibles à l\'écoute et au téléchargement au format mp3 <a href=\"https://www.enpleineconscience.ch/meditations-guidees-mp3/\">sur mon site</a>.</p>',
+    slug: 'silence',
+    description: 'Cet enregistrement de silence avec des sons de bol permet de méditer sans instructions. 1 \"ding\" après 10 minutes, 2 après 20, 3 après 30, etc. jusqu\'à 6 coups pour 60 minutes.',
     recordings: [
         {
             name: '(mbsr) silence avec \"dings\" (de 10 à 60 min)',
