+++ src/config/categories/talks.js
@@ -1,6 +1,7 @@
 export default {
     name: 'causeries',
-    description: 'Quelques extraits des causeries données durant des retraites ou d\'autres formations réservées à mes participants. <p>Vous trouverez aussi des causeries dans mon podcast, dans un format plus adapté aux longues causeries:</p> <a href=\"https://www.enpleineconscience.ch/?feed=podcast\">Podcast (lien universel RSS)</a><br><a href=\"https://itunes.apple.com/ch/podcast/les-causeries-denpleineconscience-ch/id1457167644?l=fr\">Podcast (iTunes)</a><br>  <h3> remarque importante </h3> <p> Pour une écoute sans interruption, je vous recommande fortement de procéder comme suit: 1) télécharger l\'enregistrement choisi à l\'aide du bouton à droite de celui-ci. 2) mettre votre appareil en mode avion 3) appuyer sur play! En cas de problème, les enregistrements sont aussi disponibles à l\'écoute et au téléchargement au format mp3 <a href=\"https://www.enpleineconscience.ch/meditations-guidees-mp3/\">sur mon site</a>. </p>',
+    slug: 'causeries',
+    description: 'Quelques extraits des causeries données durant des retraites ou d\'autres formations réservées à mes participants. <p>Vous trouverez aussi des causeries dans mon podcast, dans un format plus adapté aux longues causeries:</p> <a href=\"https://www.enpleineconscience.ch/?feed=podcast\">Podcast (lien universel RSS)</a><br><a href=\"https://itunes.apple.com/ch/podcast/les-causeries-denpleineconscience-ch/id1457167644?l=fr\">Podcast (iTunes)</a>',
     recordings: [
         {
             name: '\"Tout passe\", 8\'',
