+++ src/index.js
@@ -1,3 +1,4 @@
+import './index.css'
 import { createApp } from 'vue';
 import { createRouter, createWebHistory } from 'vue-router';
 import App from './index.vue';
