+++ src/index.vue
@@ -1,11 +1,24 @@
+<!--
+    Template
+-->
+
 <template>
-  <div>
-    <router-view></router-view>
-  </div>
+    <l-header />
+    <main class="wrapper mt-20 mb-24 py-8">
+      <router-view />
+    </main>
+    <l-download />
 </template>
 
-<script>
-export default {
-  name: 'App'
-};
+
+
+<!--
+    Script
+-->
+
+<script setup>
+
+  import LHeader from '#src/components/l-header.vue'
+  import LDownload from '#src/components/l-download.vue'
+
 </script>
